add: port storypoints to typescript
